Vehicle fixes
After a lengthly discussion and some research on the part of QA I have changed vehicles so that they spawn to the side of the player

When a player becomes incapacitated or knocked down while mounted they will now be dismounted as part of being knocked down and/or incapacitated.

The default action for swoops should once again be to mount/dismount... this means you can double click the speeder to mount or dismount. This comes @ a cost and that is that the radials will now be misnumbered (1,1,2) so we are trading a fix for a lesser bug >.<.