trivial: changes to heartbeat log messages to show correct # players on zone and to show # clients login has handled.
addressed array out of bounds issue on resource list on frog: http://paste.swganh.org/viewp.php?id=20100626223514-c5382a0920ce58ada331f5d3b6828c66

added a very basic 'send zone message' ability for server admins to use from the zone window while running. works very similar to 'l' and 'q' (lock, quit). syntax: m <message here>  (ensure return follows). Use at your own peril...

added a utility method to our HeightMap that allows passing of 3 float values. One should be the heightmaps determination of a point, one is the value provided by the client, and the third is an allowed variance between the two. This is temporary until we can fix heightmaps themselves and not have to double-check ourselves like this. Updated vehicle calls (generation) and structure placement to use the compensation method with a variance of 10.0. This is an improvement to an existing change that did the same (albeit in a different way) for vehicles only.